fix lucky gift

+++ src/test/java/vnpay/vietcombank/lucky_gift/LuckyGift.java
@@ -52,7 +52,7 @@ public class LuckyGift extends Base {
 	}
 
 	@Parameters({ "pass" })
-	//@Test
+	// @Test
 	public void TC_01_NGuoiNhanTrongVCBBangSDTXacThucBangMatKhau(String pass) {
 		log.info("TC_01_Step_1: Chọn quà tặng may mắn");
 		luckyGift.clickToDynamicButtonLinkOrLinkText(driver, LuckyGift_Data.TitleLuckyGift.TITLE);
@@ -64,7 +64,7 @@ public class LuckyGift extends Base {
 		String moneySurplus = luckyGift.getDynamicTextDetailByIDOrPopup(driver, "com.VCB:id/available_balance");
 		String[] moneySurplusSplit = moneySurplus.split(" ");
 		String moneySurplusReplace = moneySurplusSplit[0].replace(",", "");
-		int surplus = Integer.parseInt(moneySurplusReplace);
+		long surplus = Integer.parseInt(moneySurplusReplace);
 
 		log.info("TC_01_Step_3: Thêm người nhận");
 		luckyGift.clickToDynamicImageViewByID(driver, "com.VCB:id/ivAdd");
@@ -107,8 +107,8 @@ public class LuckyGift extends Base {
 		log.info("TC_01_Step_14: lấy ra phí chuyển");
 		moneyFee = luckyGift.getTextInDynamicDropdownOrDateTimePicker(driver, "com.VCB:id/tvSumfee");
 		String[] sumFee = moneyFee.split(" ");
-		int sumFeeInt = Integer.parseInt(sumFee[0].replace(",", ""));
-		int surplusTotal = surplus - Integer.parseInt(LuckyGift_Data.LuckyGift.MONEY) - sumFeeInt;
+		long sumFeeInt = Integer.parseInt(sumFee[0].replace(",", ""));
+		long surplusTotal = surplus - Integer.parseInt(LuckyGift_Data.LuckyGift.MONEY) - sumFeeInt;
 		surplusString = String.valueOf(surplusTotal);
 
 		log.info("TC_01_Step_15: Click tiep tuc popup");
@@ -159,7 +159,7 @@ public class LuckyGift extends Base {
 
 	}
 
-	//@Test
+	// @Test
 	public void TC_02_BaoCaoTienQuaTangMayManTrongVCBBangSDTVaXacThucBangMK() {
 		log.info("TC_02_Step_02: Click vao More Icon");
 		homePage.clickToDynamicImageViewByID(driver, "com.VCB:id/menu_5");
@@ -228,7 +228,7 @@ public class LuckyGift extends Base {
 	}
 
 	@Parameters({ "pass" })
-	//@Test
+	// @Test
 	public void TC_03_NGuoiNhanTrongVCBBangSTKXacThucBangMatKhau(String pass) {
 		log.info("TC_03_Step_1: Chọn quà tặng may mắn");
 		luckyGift.clickToDynamicButtonLinkOrLinkText(driver, LuckyGift_Data.TitleLuckyGift.TITLE);
@@ -327,7 +327,7 @@ public class LuckyGift extends Base {
 
 	}
 
-	//@Test
+	// @Test
 	public void TC_04_BaoCaoTienQuaTangMayManTrongVCBBangSTKVaXacThucBangMK() {
 		log.info("TC_04_Step_02: Click vao More Icon");
 		homePage.clickToDynamicImageViewByID(driver, "com.VCB:id/menu_5");
@@ -406,7 +406,7 @@ public class LuckyGift extends Base {
 		String moneySurplus = luckyGift.getDynamicTextDetailByIDOrPopup(driver, "com.VCB:id/available_balance");
 		String[] moneySurplusSplit = moneySurplus.split(" ");
 		String moneySurplusReplace = moneySurplusSplit[0].replace(",", "");
-		int surplus = Integer.parseInt(moneySurplusReplace);
+		long surplus = Integer.parseInt(moneySurplusReplace);
 
 		log.info("TC_05_Step_3: chọn hình thức gửi quà");
 		luckyGift.clickToTextID(driver, "com.VCB:id/tvHinhThuc");
@@ -454,8 +454,8 @@ public class LuckyGift extends Base {
 		log.info("TC_05_Step_15: lấy ra phí chuyển");
 		String moneyFee = luckyGift.getTextInDynamicDropdownOrDateTimePicker(driver, "com.VCB:id/tvSumfee");
 		String[] sumFee = moneyFee.split(" ");
-		int sumFeeInt = Integer.parseInt(sumFee[0].replace(",", ""));
-		int surplusTotal = surplus - Integer.parseInt(LuckyGift_Data.LuckyGift.MONEY) - sumFeeInt;
+		long sumFeeInt = Integer.parseInt(sumFee[0].replace(",", ""));
+		long surplusTotal = surplus - Integer.parseInt(LuckyGift_Data.LuckyGift.MONEY) - sumFeeInt;
 		surplusString = String.valueOf(surplusTotal);
 
 		log.info("TC_05_Step_16: Click tiep tuc popup");
@@ -580,7 +580,7 @@ public class LuckyGift extends Base {
 		String moneySurplus = luckyGift.getDynamicTextDetailByIDOrPopup(driver, "com.VCB:id/available_balance");
 		String[] moneySurplusSplit = moneySurplus.split(" ");
 		String moneySurplusReplace = moneySurplusSplit[0].replace(",", "");
-		int surplus = Integer.parseInt(moneySurplusReplace);
+		long surplus = Integer.parseInt(moneySurplusReplace);
 
 		log.info("TC_07_Step_3: Thêm người nhận");
 		luckyGift.clickToDynamicImageViewByID(driver, "com.VCB:id/ivAdd");
@@ -619,8 +619,8 @@ public class LuckyGift extends Base {
 		log.info("TC_07_Step_13: lấy ra phí chuyển");
 		String moneyFee = luckyGift.getTextInDynamicDropdownOrDateTimePicker(driver, "com.VCB:id/tvSumfee");
 		String[] sumFee = moneyFee.split(" ");
-		int sumFeeInt = Integer.parseInt(sumFee[0].replace(",", ""));
-		int surplusTotal = surplus - Integer.parseInt(LuckyGift_Data.LuckyGift.MONEY) - sumFeeInt;
+		long sumFeeInt = Integer.parseInt(sumFee[0].replace(",", ""));
+		long surplusTotal = surplus - Integer.parseInt(LuckyGift_Data.LuckyGift.MONEY) - sumFeeInt;
 		surplusString = String.valueOf(surplusTotal);
 
 		log.info("TC_07_Step_14: chọn phương thức xác thực");
@@ -757,7 +757,7 @@ public class LuckyGift extends Base {
 		String moneySurplus = luckyGift.getDynamicTextDetailByIDOrPopup(driver, "com.VCB:id/available_balance");
 		String[] moneySurplusSplit = moneySurplus.split(" ");
 		String moneySurplusReplace = moneySurplusSplit[0].replace(",", "");
-		int surplus = Integer.parseInt(moneySurplusReplace);
+		long surplus = Integer.parseInt(moneySurplusReplace);
 
 		log.info("TC_09_Step_3: Thêm người nhận");
 		luckyGift.clickToDynamicImageViewByID(driver, "com.VCB:id/ivAdd");
@@ -797,8 +797,8 @@ public class LuckyGift extends Base {
 		log.info("TC_09_Step_14: lấy ra phí chuyển");
 		String moneyFee = luckyGift.getTextInDynamicDropdownOrDateTimePicker(driver, "com.VCB:id/tvSumfee");
 		String[] sumFee = moneyFee.split(" ");
-		int sumFeeInt = Integer.parseInt(sumFee[0].replace(",", ""));
-		int surplusTotal = surplus - Integer.parseInt(LuckyGift_Data.LuckyGift.MONEY) - sumFeeInt;
+		long sumFeeInt = Integer.parseInt(sumFee[0].replace(",", ""));
+		long surplusTotal = surplus - Integer.parseInt(LuckyGift_Data.LuckyGift.MONEY) - sumFeeInt;
 		surplusString = String.valueOf(surplusTotal);
 
 		log.info("TC_09_Step_15: chọn phương thức xác thực OTP");
@@ -936,7 +936,7 @@ public class LuckyGift extends Base {
 		String moneySurplus = luckyGift.getDynamicTextDetailByIDOrPopup(driver, "com.VCB:id/available_balance");
 		String[] moneySurplusSplit = moneySurplus.split(" ");
 		String moneySurplusReplace = moneySurplusSplit[0].replace(",", "");
-		int surplus = Integer.parseInt(moneySurplusReplace);
+		long surplus = Integer.parseInt(moneySurplusReplace);
 
 		log.info("TC_11_Step_3: chọn hình thức gửi quà");
 		luckyGift.clickToTextID(driver, "com.VCB:id/tvHinhThuc");
@@ -984,8 +984,8 @@ public class LuckyGift extends Base {
 		log.info("TC_11_Step_15: lấy ra phí chuyển");
 		String moneyFee = luckyGift.getTextInDynamicDropdownOrDateTimePicker(driver, "com.VCB:id/tvSumfee");
 		String[] sumFee = moneyFee.split(" ");
-		int sumFeeInt = Integer.parseInt(sumFee[0].replace(",", ""));
-		int surplusTotal = surplus - Integer.parseInt(LuckyGift_Data.LuckyGift.MONEY) - sumFeeInt;
+		long sumFeeInt = Integer.parseInt(sumFee[0].replace(",", ""));
+		long surplusTotal = surplus - Integer.parseInt(LuckyGift_Data.LuckyGift.MONEY) - sumFeeInt;
 		surplusString = String.valueOf(surplusTotal);
 
 		log.info("TC_11_Step_16: chọn phương thức xác thực OTP");
